pre-feature cleanup: docs, imports, and vector storage validation extraction
A collection of small, self-contained cleanups that don't depend on
the NormalizedVector type. Pulled out so they can land (and be
reviewed) independently of the NormalizedVector feature.

- vortex-array: relax ExtensionArray storage nullability check to
  allow the storage array's outer nullability to differ from the
  extension's declared storage outer nullability. Nested storage
  nullability still must match exactly.
- vortex-tensor/types/vector: extract storage-dtype validation out
  of the Vector vtable into a crate-visible
  `validate_vector_storage_dtype` helper in `vector/mod.rs`.
- vortex-tensor/utils: add `use half::f16;` and simplify
  `as_slice::<half::f16>()` to `as_slice::<f16>()` in `cast_to_f32`.
- vortex-tensor/scalar_fns/l2_norm: fix intra-doc link path for
  `L2Denorm` in the struct docs; add an "Inherit the nullability
  from the vectors themselves." explanatory comment in
  `return_dtype`.
- vortex-tensor/encodings/l2_denorm: add an explanatory comment on
  `expected_compression_ratio` noting that pre-normalization is
  almost always desired.
- vortex-tensor/scalar_fns/inner_product: reword the
  `BinaryTensorOpMetadata::decode_children` doc comment.
- vortex-tensor/encodings/turboquant/compress: rewrap the
  `turboquant_encode` "Errors" doc bullet.

+++ vortex-tensor/src/utils.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+use half::f16;
 use vortex_array::ArrayRef;
 use vortex_array::ExecutionCtx;
 use vortex_array::IntoArray;
@@ -73,7 +74,7 @@ pub fn validate_binary_tensor_float_inputs<'a>(
 pub fn cast_to_f32(prim: PrimitiveArray) -> VortexResult<Buffer<f32>> {
     match prim.ptype() {
         PType::F16 => Ok(prim
-            .as_slice::<half::f16>()
+            .as_slice::<f16>()
             .iter()
             .map(|&v| f32::from(v))
             .collect()),
